makefile will re-create zipped artifacts
To save in git we unzip the .mtmacro, .mtmacset, and .rptok
to constituent parts.  This makefile trivially reconsitutes
those.

To use:

make <dir>.<thing>

<dir>
    should be a directory with content.xml and properties.xml
<thing>
    should be one of mtmacro mtmacset or rptok

    There is no functional difference, but the exported items had
    different names so we're going to put them back that way.

Testing may be needed to make sure we know if the filename
matters, and whether contents.xml may contain the _type_ of
object -- in which case we can try to read it from there.