xc and automagic now in docker/maker
Reoriented the two docker images so that maker-dev is based
on maker instead of the other way around.  Let me know if
that's not kosher.   A side-effect of this is that automagic
and xc are available in maker-dev.  The /usr/local/bin copy
and the local filesystem copy.

the dockerbuild and makefile no longer have to concern
themselves about moving and removing automagic and xc.  They
live in the place docker expects them now.