micro-assemble to reassemble a macro
does not handle rptok, mtmacset, or mtprops

you have to specify the two input parts and the output part
name on the commandline, this one doesn't do any of that for
you.

But, it results in a .mtmacro file (should you name it that)
that will import and run the contents of the command
filename as a macro.