Issue #5 file not found message improvement
I kept wanting to run this via docker on a file that was not in
the current directory (what do you mean my file is not where it
obviously is?).  So have it printing a hint so I'll hopefully
stop trying the same thing.

It turns out you can't start your commit message with _only_ #5,
needs a lead-in character that isn't #