Issue #20 repo layout + project file spec
The Lib: tokens' macros will now be extracted with them into
their own directory, illustrated here.

The macro/ will be used for extracting .mtmacro, .mtmacset, and
.rptok macros where the macro name does not start with 'Lib:'.

A project can be specified to have a bunch of parts as in
DNDBeyond.project.  Note no work yet to parse that into actions.