Issue #20 reorganize basically complete
removed extra directories that just have macros, after ensuring
their macro contents are in macro/

found the real name of things we intended for DNDBeyond.project

Removed 'xc' as we won't be using that trivial extractor, the
macro-extract is all working nicely now.

Extracted the DNDBeyond-mtmacset to macro/ and left a breadcrumb
in DNDBeyond.project.   Will remove the superfluous mtmacset
directory one of these next few go-arounds.  Did the same with
the Roll Initiative macset, but I went ahead and cleaned those
up.