Issue #5 Break out the rptok contents
This is the exploded rptok contents

rptok (zip)
  content.xml
    macroPropertiesMap
      entry  [A]
        net.rptools.maptool.model.MacroButtonProperties [B]
          command [C]

Each entry [A] in content.xml will be replaced with a placeholder
<entry>. It will have the same <int> member, but will have it's macro
replaced with a simple <macro name="label"/> placeholder.

The label will be used to find the macros on the filesystem.

label.xml will contain [B] except for the <command></command> (it will
not be present).  label.command will contain the text of [C].

In this way the repository will show all these objects separately.

These are as-built by token-extract.   token-assemble puts them all back
together into the rptok.

token-extract takes the .rptok as its argument (--verbose to see more)
% ./dockerrun token-extract Lib-DND5e.rptok

token-assemble takes the directory as its argument (--verbose to see
more)
% ./dockerrun token-assemble Lib-DND5e