Issue #41 README/automagic/makefile all agree
The README was still saying automagic, which was still using the
makefile, which was a previous style of rebuilding the assets.  This was
not compatible with the new macro-extract and token-extract.

README now explains this and automagic calls macro-assemble or
token-assemble if called on a macro or token.   automagic cannot
assemble macsets, those are done right now by calling macro-assemble
with the list of macros and --set SetName.