Add GM/Owner Health Bar
Added a health bar that represents the % of Max HP of the tokens current HP. Also added a bar that represents the % of Max HP for the current temporary HP that goes on top of the current HP bar. Only the GM and owner of the token can see the bars.
Updated dndb_applyHealth to show these bars