Use make dependencies to build images
'make' does a good job of tracking dependencies when we let it, so I've
created a file target to track the timestamp the last images were built.

Make only needs to build when files in the docker directory change, and
the makefile now connects the dots.

The dockerbuild script is still valid to build both images.