Fixes build a .project file
This introduces project-assemble script for running inside docker to
assemble all the "things" listed in a project xml file.  The project xml
file needed to be caught up with the names of things on the filesystem.

Currently, the default verbosity is different based on which thing is
doing the assembly.  Minor annoyance.  But a little bigger is that you
can't change them up nor down.

Trying a new format in the README.md per the github documentation, but I
don't have a github markdown renderer, so if it sucks I'll change it
again.