Fixes empty command on token-extract
token-extract had never seen a macro with an empty command so was
dealing with it poorly.  This will write a zero length file for empty
commands so that they're complete, and when putting it back together the
command is empty.

note that when we reassemble empty elements, they do come out slightly
different with <command></command> represented as <command/>.  I have
not found a way to tell the library to prefer the former yet.  These
were tested in maptool and were observed to not have a negative impact.