Fix output via logging
With the reorganization through a library, the output wasn't working as
expected.  Using the logging module lets me control the output
throughout.

Adding a newline to Lib-DND5e/properties.xml (the new writer adds it).

This starts the work to standardize the version string in properties.xml
to the version in MTAssetLibrary.  Not completed yet.