Added getInvestigation and fixed Initiative
Initiative now rolls the same as - separate, but doesn't initialize the
init pane.   It happens more frequently than the player like to have
something join the initiative.

I also wanted to see the Investigation state on token sheets, so I added
it.  This was added first to git, then I built the assets and imported
them.  Worked well for these smaller changes that could be based on an
existing one.