Issue #68 Introduce BDD, start output fix
This introduces BDD to test interfaces for the building tools.   BDD
feature descriptions thus far are in test/features/assemble.feature.

BDD is run through the behave docker image which will automatically
run the tests.   The makefile can launch the test on docker with
'make behave'.

As part of this begun refactoring of the assemble (and other tools)
with an asset object doing the lifting.  The interface files will be
like token-assemble now, a relatively thin veneer.