Attack output update
Added a lot of support to existing DiceRoller commands to save data needed when generating output. This data is stored in a typeDescriptors object on the RollExpression. New methods were added to create and update this object.
Created a new formatter that shows the attack rolls with less text and more focus on the actual results. The details are added to the tool tips for each roll.