Added health AUDIT log message
The message puts the current and new values for HP, MaxHP & TempHP in a log message. The message starts with the text 'AUDIT TOKEN HEALTH' so that it can be easy to find the value if needed during a game.

All other dnd5e health related commands have had their existing log messages set to the debug level.