Issue #68 extracted token and mtasset testing
This creates an extracted token that is in a plain zip for the assemble
tests to use.  This with the Minimum Viable candidates are used in the
current test suite(s).

test_mtasset added to check the behavior of our MTAsset object when
assembling.

test_extract added to check the behavior of our MTAsset object when
extracting (WIP).