Attack results can now be applied to tokens
After an attack from the attack editor is executed the attack is saved on the DnD5e library token. Then that attack can be applied to as many tokens as needed.

While applying the attack immunities, resistances and vulnerabilities are applied to the damage rolls.

It will also automatically make saving throws for damage and apply effects if the save is successful.

Added the Apply Last Attack macro button to the campaign macro set, but this macro can only ran by the GM. All of its ouput is GM only information.