Changes to restore dead NPC
When an NPC token dies and is sent to the object layer the token's ID and current initiative are saved in a list on the "_deadTokens" Lib:DnD5e token property. This list will remember the last ten dead tokens. Adding another token to the list after it already has 10 removes the oldest token in the list.
A new macro has been created that reads the last token to be added to the list, heals it for a hit point, moves it to the token layer, adds it to initiative, then sets it initiative value to be the same value as when it died. THen it sorts the initiative list to get that token in the right spot.