Issue #68 assemble complete
This version of assemble is complete.

It combines the multiple versions for different input types into
one script and one set of libraries doing the lifting.   The assemble
command has symlinks for the previous command interfaces, which work
the same.

To change the directory where the output is assembled we have
introduced an --output argument and a config file:

config.ini:
[default]
outputdirectory = output

This will put all the assembled assets into 'output' by default.