Added new action editor
This provides a new UI and enhanced functionality for the old Attack Editor. It supports setting states. It adds fields to support applying a roll to a target. Extends roll expression to support expressions there to help with conditions and reading data from DNDBeyond. It supports the spell casting macros by converting their output into actions.