Added field tooltips and validation
Mostly validation meant that most fields were required. Extended fields and Save Against are not required. Roll field uses a regex to validate. Bonus and DC fields require a number.

Made many of the buttons not perform validation. Includes the delete and extendsion buttons, all of the add step buttons, the Delete button and the Exit button.

It is no longer possible to delete the last step in an action.

Made the DnDBeyond Spell, Spell Level, and Weapon fields automatically submit the form so that their related changes can be applied.

Added tooltips to all fields.