all target now makes a DNDBeyond+Open5e project
DNDBeyond+Open5e.project is a merge of DNDBeyond.project and
Open5e.project.

The asset.py was throwing an error on creating 'output' which it
shouldn't have, so I've swallowed that error for now.

The makefile targets 'project' (default) and 'all' have been updated to
run the assemble tool instead of project-assemble (which is just a
symlink to assemble).

Returned the Open5e.project to its former state, as it is "standalone"
and needs the GM Preferences macro (and the Set Default Map is icing).