Issue #164 refactoring complete for tokens
reading in token artifacts from the filesystem results in a new
MTToken object which knows how to .assemble() and .extract() itself.

macro artifacts are close behind with MTMacroObj also gaining .append()
to take a macro and append another macro, converting to a macroset.

macrosets objects will get the same treatment, all on a learning curve
toward being able to append/convert contents of properties (tieing it
all back to #164).