Issue #176 #177 #183 refactoring feature complete
This push has all the changes from

Issue #176 - embed instructions for the GM into project files

    projects now take a <text> with an optional name attribute
    and write everythign between <text> and </text> to a file.
    if name is not specified it will use README.txt

Issue #177 - make project files reference another project file

    projects now take a <project name="Foo"/> or <project
    name="Foo.project/> tag and will also build that project.

Issue #183 - insert commit hash into token/macro/mtprops content.xml

    the git hash is written into token <gmName> as specified.  Currently
    the comment at the top of the token is a little verbose, working
    on simplifying that.  Haven't yet applied it to properties.

This is all in preparation to actually go after #164 (on the dev-164
branch) and figure out how to append things to properties lists.

added many tests with this refactor.  The python was tested with 16
tests, now 47 so hopefully additional coverage in addition to tests
for the new features.

This commit is not _yet_ behave complete, so there are some interface
issues to deal with.