CampaignMerge working, integration incomplete
This version of CampaignMerge puts the merged xml to stdout.  According
to lore, one should just let the errors bubble up, so this does, but
I feel like it should print something useful instead of a trace (trace
says what went wrong, not what the operator should change to make it
work right).