fixes #272 tokens assemble macros from files
This changes the assembly of tokens to pick up all .xml files that
are not otherwise known (content.xml, properties.xml, and
propertyMapCI.xml) and load them as macros.

The macroPropertiesMap from extract is still in the content.xml on
the filesystem, but it is no longer read (it is thrown away when loading
the asset from the filesystem).